+++ disentangled_rnns/library/rnn_utils_test.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import json
-
 from absl.testing import absltest
 from disentangled_rnns.library import rnn_utils
 from disentangled_rnns.library import two_armed_bandits
@@ -26,7 +25,7 @@ import optax
 
 n_steps_per_session = 10
 n_sessions = 3
-batch_size = n_sessions
+batch_size = 1024
 n_hidden = 20
 
 
@@ -51,7 +50,7 @@ class TestRNNUtils(absltest.TestCase):
         environment,
         n_steps_per_session=n_steps_per_session,
         n_sessions=n_sessions,
-        batch_size=batch_size)
+        )
 
     # "Train" for zero steps to instantiate model variables
     self.params, self.opt_state, _ = rnn_utils.train_network(
@@ -64,7 +63,8 @@ class TestRNNUtils(absltest.TestCase):
         opt=optax.adam(learning_rate=0.001))
 
   def test_dataset_rnn(self):
-    xs, ys = next(self.dataset)
+    data = next(self.dataset)
+    xs, ys = data['xs'], data['ys']
 
     # Inputs for y-maze should be choice & reward
     self.assertEqual(np.shape(xs), (n_steps_per_session, batch_size, 2))
@@ -79,9 +79,9 @@ class TestRNNUtils(absltest.TestCase):
         xs=np.zeros((n_steps_per_session, n_sessions, 2)),
         ys=np.zeros((n_steps_per_session, n_sessions, 1)),
         y_type='categorical',
-        batch_size=None)
-    xs, _ = next(dataset)
-    self.assertEqual(np.shape(xs), (n_steps_per_session, n_sessions, 2))
+        batch_size=10)
+    xs = next(dataset)['xs']
+    self.assertEqual(np.shape(xs), (n_steps_per_session, 10, 2))
 
   def test_dataset_rnn_rolling_batch_gt_nepisodes(self):
     """Test rolling batch mode when batch_size > n_episodes."""
@@ -106,14 +106,14 @@ class TestRNNUtils(absltest.TestCase):
 
     # Initial order: [0, 1, 2]
     # First batch expected indices: tile([0,1,2], 2)[:5] = [0, 1, 2, 0, 1]
-    xs_batch1, _ = next(dataset)
+    xs_batch1 = next(dataset)['xs']
     self.assertEqual(xs_batch1.shape, (n_timesteps, batch_s, n_features))
     np.testing.assert_array_equal(xs_batch1[0, :, 0], [0, 1, 2, 0, 1])
     # Expected start index after 1st batch: (0 + 5) % 3 = 2
     self.assertEqual(dataset._current_start_index, 2)
 
     # Second batch expected indices: tile([2,0,1], 2)[:5] = [2, 0, 1, 2, 0]
-    xs_batch2, _ = next(dataset)
+    xs_batch2 = next(dataset)['xs']
     self.assertEqual(xs_batch2.shape, (n_timesteps, batch_s, n_features))
     np.testing.assert_array_equal(xs_batch2[0, :, 0], [2, 0, 1, 2, 0])
     # Expected start index after 2nd batch: (2 + 5) % 3 = 1
@@ -140,14 +140,14 @@ class TestRNNUtils(absltest.TestCase):
     )
     # Initial order: [0, 1, 2]
     # First batch: [0, 1]
-    xs_batch1, _ = next(dataset)
+    xs_batch1 = next(dataset)['xs']
     self.assertEqual(xs_batch1.shape, (n_timesteps, batch_s, n_features))
     np.testing.assert_array_equal(xs_batch1[0, :, 0], [0, 1])
     # Expected start index after 1st batch: (0 + 2) % 3 = 2
     self.assertEqual(dataset._current_start_index, 2)
 
     # Second batch: [2, 0]
-    xs_batch2, _ = next(dataset)
+    xs_batch2 = next(dataset)['xs']
     np.testing.assert_array_equal(xs_batch2[0, :, 0], [2, 0])
     # Expected start index after 2nd batch: (2 + 2) % 3 = 1
     self.assertEqual(dataset._current_start_index, 1)
@@ -173,24 +173,32 @@ class TestRNNUtils(absltest.TestCase):
     )
     # Initial order: [0, 1, 2]
     # First batch: [0, 1, 2]
-    xs_batch1, _ = next(dataset)
+    xs_batch1 = next(dataset)['xs']
     self.assertEqual(xs_batch1.shape, (n_timesteps, batch_s, n_features))
     np.testing.assert_array_equal(xs_batch1[0, :, 0], [0, 1, 2])
     # Expected start index after 1st batch: (0 + 3) % 3 = 0
     self.assertEqual(dataset._current_start_index, 0)
 
     # Second batch: [0, 1, 2]
-    xs_batch2, _ = next(dataset)
+    xs_batch2 = next(dataset)['xs']
     np.testing.assert_array_equal(xs_batch2[0, :, 0], [0, 1, 2])
     # Expected start index after 2nd batch: (0 + 3) % 3 = 0
     self.assertEqual(dataset._current_start_index, 0)
 
   def test_split_dataset(self):
     dataset_train, dataset_eval = rnn_utils.split_dataset(self.dataset, 2)
-    xs_train, _ = next(dataset_train)
-    xs_eval, _ = next(dataset_eval)
-    self.assertEqual(np.shape(xs_train), (n_steps_per_session, 2, 2))
-    self.assertEqual(np.shape(xs_eval), (n_steps_per_session, 1, 2))
+    xs_train_all = dataset_train.get_all()['xs']
+    xs_eval_all = dataset_eval.get_all()['xs']
+    self.assertEqual(np.shape(xs_train_all), (n_steps_per_session, 2, 2))
+    self.assertEqual(np.shape(xs_eval_all), (n_steps_per_session, 1, 2))
+    xs_train_batch = next(dataset_train)['xs']
+    xs_eval_batch = next(dataset_eval)['xs']
+    self.assertEqual(
+        np.shape(xs_train_batch), (n_steps_per_session, batch_size, 2)
+    )
+    self.assertEqual(
+        np.shape(xs_eval_batch), (n_steps_per_session, batch_size, 2)
+    )
 
   def test_train_network(self):
     """Train the network for a few steps, check that the loss goes down.
@@ -216,16 +224,79 @@ class TestRNNUtils(absltest.TestCase):
     # Check that opt state has changed
     self.assertNotEqual(self.opt_state, new_opt_state)
 
+  def test_get_new_params(self):
+    """Test that get_new_params returns a new set of params."""
+    xs = next(self.dataset)['xs']
+    input_size = xs.shape[-1]
+    new_params = rnn_utils.get_new_params(self.make_network,
+                                          input_size=input_size)
+    self.assertNotEmpty(new_params, 'new_params should not be empty')
+    self.assertIn('gru', new_params)
+
+  def test_training_from_new_params(self):
+    """Test that training from new params works."""
+    xs = next(self.dataset)['xs']
+    input_size = xs.shape[-1]
+    new_params = rnn_utils.get_new_params(self.make_network,
+                                          input_size=input_size)
+
+    _, _, losses = rnn_utils.train_network(
+        self.make_network,
+        training_dataset=self.dataset,
+        validation_dataset=self.dataset,
+        random_key=self.random_key,
+        loss='categorical',
+        n_steps=100,
+        opt=optax.adam(learning_rate=0.01),
+        opt_state=None,
+        params=new_params)
+    self.assertGreater(losses['training_loss'][0],
+                       losses['training_loss'][-1])
+
+  def test_train_network_from_json_params(self):
+    """Smoke test for training from params loaded from json."""
+
+    # Generate params, convert them to json, convert them back
+    initial_params, _, _ = rnn_utils.train_network(
+        self.make_network,
+        validation_dataset=None,
+        training_dataset=self.dataset,
+        loss='categorical',
+        n_steps=0,
+    )
+    json_string = json.dumps(initial_params, cls=rnn_utils.NpJnpJsonEncoder)
+    loaded_params = json.loads(json_string)
+
+    # Check that we're able to train from loaded params
+    new_params, _, losses = rnn_utils.train_network(
+        self.make_network,
+        training_dataset=self.dataset,
+        validation_dataset=self.dataset,
+        random_key=self.random_key,
+        loss='categorical',
+        n_steps=10,  # A small number of steps for a smoke test
+        opt=optax.adam(learning_rate=0.01),
+        params=loaded_params,
+    )
+
+    self.assertGreater(losses['training_loss'][0], losses['training_loss'][-1])
+    initial_params_after_conversion = rnn_utils.to_np(loaded_params)
+    self.assertFalse(
+        np.all(
+            initial_params_after_conversion['gru']['w_h']
+            == new_params['gru']['w_h']
+        )
+    )
+
   def test_eval_network(self):
-    """Eval a network on a set of inputs. Check shapes look right.
-    """
+    """Eval a network on a set of inputs. Check shapes look right."""
     # Get a set of inputs
-    xs, _ = next(self.dataset)
+    xs = self.dataset.get_all()['xs']
     # Eval the network on that set of inputs
     y_hats, states = rnn_utils.eval_network(make_network, self.params, xs)
 
     # Check outputs have the right shape
-    self.assertEqual(np.shape(y_hats), (n_steps_per_session, batch_size, 2))
+    self.assertEqual(np.shape(y_hats), (n_steps_per_session, n_sessions, 2))
     # Check states have the right shape
     self.assertEqual(
         np.shape(states), (n_steps_per_session, n_sessions, n_hidden))
@@ -240,7 +311,7 @@ class TestRNNUtils(absltest.TestCase):
     """Step the network forward two steps. Check the shapes all look right.
     """
     # Get a set of inputs
-    xs, _ = next(self.dataset)
+    xs = next(self.dataset)['xs']
     # Define a network state
     state = rnn_utils.get_initial_state(self.make_network)